refactor: FSD 구조로 초기 변경 완료 (#99)
