+++ apps/web/src/app/(with-global-layout)/auth/layout.tsx
@@ -1,5 +1,5 @@
 import React, { Suspense } from 'react';
-import PageContainer from 'src/components/layout/PageContainer';
+import PageContainer from 'src/shared/ui/PageContainer';
 
 export default function AuthLayout({ children }: { children: React.ReactNode }) {
   return (
