+++ apps/web/src/app/(with-local-layout)/(with-header-layout)/mypage/page.tsx
@@ -37,3 +37,11 @@
 //   }
 // };
 // export default MyPage;
+
+import React from 'react';
+
+const page = () => {
+  return <div>page</div>;
+};
+
+export default page;
