+++ apps/web/src/app/(with-local-layout)/mypage/auctions/page.tsx
@@ -14,3 +14,11 @@
 // };
 
 // export default MyAuctions;
+
+import React from 'react';
+
+const page = () => {
+  return <div>page</div>;
+};
+
+export default page;
