+++ apps/web/src/app/(with-local-layout)/mypage/episodes/page.tsx
@@ -17,3 +17,11 @@
 // };
 
 // export default MyStories;
+
+import React from 'react';
+
+const page = () => {
+  return <div>page</div>;
+};
+
+export default page;
