+++ apps/web/src/app/(with-local-layout)/mypage/points/page.tsx
@@ -39,3 +39,11 @@
 // };
 
 // export default MyPoints;
+
+import React from 'react';
+
+const page = () => {
+  return <div>page</div>;
+};
+
+export default page;
