+++ apps/web/src/app/api/auth/logout/route.ts
@@ -1,14 +1,14 @@
-import { NextResponse } from 'next/server';
-import { getAuthLogout } from 'src/entities/auth/supabase';
+// import { NextResponse } from 'next/server';
+// import { getAuthLogout } from 'src/entities/auth/supabase';
 
-export async function POST() {
-  try {
-    const res = await getAuthLogout();
-    return NextResponse.json({ status: 'success', data: res });
-  } catch (error) {
-    if (error instanceof Error) {
-      return NextResponse.json({ status: 'error', error: 'Server Error: ' + error.message }, { status: 500 });
-    }
-    return NextResponse.json({ status: 'error', error: 'Server Error: Unknown error' }, { status: 500 });
-  }
-}
+// export async function POST() {
+//   try {
+//     const res = await getAuthLogout();
+//     return NextResponse.json({ status: 'success', data: res });
+//   } catch (error) {
+//     if (error instanceof Error) {
+//       return NextResponse.json({ status: 'error', error: 'Server Error: ' + error.message }, { status: 500 });
+//     }
+//     return NextResponse.json({ status: 'error', error: 'Server Error: Unknown error' }, { status: 500 });
+//   }
+// }
