+++ apps/web/src/app/api/auth/user-role/route.ts
@@ -1,14 +1,14 @@
-import { NextResponse } from 'next/server';
-import { getUserRole } from 'src/shared/supabase/query/users';
+// import { NextResponse } from 'next/server';
 
-export async function GET() {
-  try {
-    const res = await getUserRole();
-    return NextResponse.json({ status: 'success', data: res });
-  } catch (error) {
-    if (error instanceof Error) {
-      return NextResponse.json({ status: 'error', error: 'Server Error: ' + error.message }, { status: 500 });
-    }
-    return NextResponse.json({ status: 'error', error: 'Server Error: Unknown error' }, { status: 500 });
-  }
-}
+// export async function GET() {
+//   try {
+
+//     const res = await getUserRole();
+//     return NextResponse.json({ status: 'success', data: res });
+//   } catch (error) {
+//     if (error instanceof Error) {
+//       return NextResponse.json({ status: 'error', error: 'Server Error: ' + error.message }, { status: 500 });
+//     }
+//     return NextResponse.json({ status: 'error', error: 'Server Error: Unknown error' }, { status: 500 });
+//   }
+// }
