+++ apps/web/src/assets/images/auctionDefault.png
