+++ apps/web/src/assets/images/noAuctionImage.png
