+++ apps/web/src/features/auction/shared/AuctionCard.tsx
@@ -12,7 +12,7 @@ import { TZDate } from 'react-day-picker';
 import { FaHeart } from 'react-icons/fa';
 import { FaBookOpen } from 'react-icons/fa6';
 import { formatNumber } from 'src/shared/utils/formatNumber';
-import noAuctionImage from '../../../assets/images/noAuctionImage.png';
+import noAuctionImage from 'src/assets/images/noAuctionImage.png';
 
 interface AuctionCardProp {
   auction_id: string;
