+++ apps/web/src/features/episode/EpisodesAuctionCard.tsx
@@ -4,7 +4,7 @@ import Image from 'next/image';
 import AuctionTimer from '../auction/AuctionTimer';
 import { selectAuctionWithSellerInfo } from 'src/entities/auction/api';
 import { AuctionRow } from 'src/shared/supabase/types';
-import NotAuctionImage from 'assets/images/auctionDefault.png';
+import NotAuctionImage from 'src/assets/images/auctionDefault.png';
 
 const EpisodesAuctionCard = async ({ auction_id }: { auction_id: AuctionRow['auction_id'] }) => {
   // NOTE - 경매 상품 및 경매 업체 정보
