+++ apps/web/src/widgets/Header.tsx
@@ -1,9 +1,9 @@
 'use client';
 import Image from 'next/image';
 import Link from 'next/link';
-import Logo from '../../../assets/images/logo.svg';
+import Logo from 'src/assets/images/logo.svg';
 import { useUserStore } from 'src/entities/auth/stores/UserStore';
-import { getAuthLogout } from 'src/entities/auth/supabase';
+// import { getAuthLogout } from 'src/entities/auth/supabase';
 import { toast } from '@repo/ui/components/ui/sonner';
 // import { IoNotifications } from 'react-icons/io5';
 // import { Button } from '@repo/ui/components/ui/button';
@@ -11,23 +11,25 @@ import { toast } from '@repo/ui/components/ui/sonner';
 const Header = () => {
   const { userInfo, setUser } = useUserStore();
 
-  const handleLogOut = async () => {
-    try {
-      const res = await getAuthLogout();
-      // 로그아웃 성공 시 store에 저장된 사용자 정보 삭제
-      if (!res.success) return;
-      setUser(null, null);
-      console.log(res.message);
-      toast.success('로그아웃 성공!');
-    } catch (error) {
-      if (error instanceof Error) {
-        console.error(error);
-      }
-    }
-  };
+  // const handleLogOut = async () => {
+  //   try {
+  //     const res = await getAuthLogout();
+  //     // 로그아웃 성공 시 store에 저장된 사용자 정보 삭제
+  //     if (!res.success) return;
+  //     setUser(null, null);
+  //     console.log(res.message);
+  //     toast.success('로그아웃 성공!');
+  //   } catch (error) {
+  //     if (error instanceof Error) {
+  //       console.error(error);
+  //     }
+  //   }
+  // };
+
+  const handleLogOut = async () => {};
 
   return (
-    <header className="h-16 w-full border-b border-b-(--color-light-gray)/30 px-5">
+    <header className="border-b-(--color-light-gray)/30 h-16 w-full border-b px-5">
       <div className="flex h-full items-center justify-between">
         <Link href="/main">
           <h1 className="text-lg font-bold">
